streamlit/requirements.txt
streamlit
openai
psycopg2-binary
pandas
plotly
python-dotenv
numpy
langsmith
mlflow
jinja2
